+++ qiskit/backends/local/statevector_simulator_py.py
@@ -19,7 +19,7 @@ The input is a qobj dictionary and the output is a Result object.
 The input qobj to this simulator has no shots, no measures, no reset, no noise.
 """
 import logging
-from qiskit._result import Result
+
 from qiskit.backends.local.localjob import LocalJob
 from qiskit.backends.local._simulatorerror import SimulatorError
 from qiskit.qobj import QobjInstruction
@@ -66,23 +66,23 @@ class StatevectorSimulatorPy(QasmSimulatorPy):
             experiment.instructions.append(
                 QobjInstruction(name='snapshot', params=[final_state_key])
             )
-        result = super()._run_job(qobj)._result
+        result = super()._run_job(qobj)
         # Replace backend name with current backend
-        result['backend'] = self._configuration['name']
+        result.backend_name = self.name
         # Extract final state snapshot and move to 'statevector' data field
-        for res in result['result']:
-            snapshots = res['data']['snapshots']
+        for experiment_result in result.results.values():
+            snapshots = experiment_result.snapshots
             if str(final_state_key) in snapshots:
                 final_state_key = str(final_state_key)
             # Pop off final snapshot added above
             final_state = snapshots.pop(final_state_key, None)
             final_state = final_state['statevector'][0]
             # Add final state to results data
-            res['data']['statevector'] = final_state
+            experiment_result.data['statevector'] = final_state
             # Remove snapshot dict if empty
             if snapshots == {}:
-                res['data'].pop('snapshots', None)
-        return Result(result)
+                experiment_result.data.pop('snapshots', None)
+        return result
 
     def _validate(self, qobj):
         """Semantic validations of the qobj which cannot be done via schemas.