+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Added
 - Add OpenMP parallelization for Apple builds of the cpp simulator (#698).
 - New interactive visualizations (#765).
 - Added option to reverse the qubit order when plotting a circuit. (#762, #786)
+- Jupyter notebook magic function qiskit_job_status (#734).
 
 Changed
 -------
@@ -36,12 +37,16 @@ Changed
     - introduce the ``qiskit.qobj`` module. (#589, #655)
     - update the ``Qobj`` JSON schema. (#668, #677, #703, #709)
     - update the local simulators for accepting ``Qobj`` as input. (#667)
+    - update the ``Result`` class. (#773)
 - Use ``get_status_job()`` for checking IBMQJob status. (#641)
 - Q network hub/group/project credentials replaced by new url format. (#740)
-- ``Jobs`` API simplification and breaking changes. (#686)
+- Breaking change: ``Jobs`` API simplification. (#686)
+- Breaking change: altered tomography APIs to not use QuantumProgram. (#818)
+- Breaking change: ``BaseBackend`` API changed, properties are now methods (#858)
 
 Removed
 -------
+- Removed the QuantumProgram class. (#724)
 
 Fixed
 -----